:lipstick: restructured the code
