+++ backend/src/model/user.ts
@@ -1,0 +1,7 @@
+import { PrismaClient } from "@prisma/client";
+
+const prisma = new PrismaClient();
+
+const User = prisma.user;
+
+export { User };
